Add files via upload
Suppose you have thought of a product or an idea for a company and you need some nice catchy title for it. Then this program would help you suggest them to you non-stop unless you are convinced by it or you want the program to terminate. This is the main concept and the prototype has achieved 50% of the main concept. The program will show any words for now which may be irrelevant to your subject. I'll be working on it in which i'll add a subject section which which will show you words related to your need and the program will be capable of using the internet as the database. For now i've a local inbuilt database for testing purpose.